fix(status): add None-safety to is_dev_collab_branch and prevent blocked_items overlap
- Make is_dev_collab_branch accept str | None with isinstance guard,
  consistent with is_orchestra_managed_flow_branch
- Exclude human_collab_numbers from blocked_items to prevent
  dev/issue-N flows appearing in both Human Collab and Blocked sections
- Split test_task_status_dashboard.py (was 690 lines) into two files:
  test_task_status_dashboard.py (375) and test_task_status_human_collab.py (414)
- Add tests for None branch input and blocked dev/issue exclusion
